Move RGB pixel's test data to separate file
