+++ test/pixels/RGBPixel.cpp
@@ -1,51 +1,15 @@
+#include "utils/RGBBasicTestData.hpp"
+
 #include <pixels/RGBPixel.hpp>
 
 #include <boost/test/unit_test.hpp>
 #include <boost/test/data/test_case.hpp>
 
-#include <array>
-#include <cstdlib>
 #include <ostream>
 
 using boost::test_tools::predicate_result;
 
-using std::array;
-using std::generate;
 using std::ostream;
-
-struct RGBPixelData
-{
-public:
-	RGBPixelData()
-	{
-		generate(red_values.begin(), red_values.end(), get_random);
-		generate(green_values.begin(), green_values.end(), get_random);
-		generate(blue_values.begin(), blue_values.end(), get_random);
-	}
-
-private:
-	static const size_t DATA_SIZE = 10;
-
-	static unsigned char get_random() noexcept
-	{
-		/*
-			* We use this to 'trim' the random number in the range of
-			* unsigned char. It is static in order not to recompute it every time.
-			*/
-		static const unsigned short over_uchar_max = UCHAR_MAX + 1;
-
-		/*
-			* We use that unsigned char is bigger than 0,
-			* as is the value returned from rand().
-			*/
-		return rand() % over_uchar_max;
-	}
-
-public:
-	array<unsigned char, DATA_SIZE> red_values;
-	array<unsigned char, DATA_SIZE> green_values;
-	array<unsigned char, DATA_SIZE> blue_values;
-};
 
 static const RGBPixelData RGB_PIXEL_TEST_DATA;
